requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.1
supabase>=2.0.0